Update js files
