+++ src/modules/nav.js
@@ -1,9 +1,12 @@
-const navBase = () => {
-  const hamburger = document.querySelector('.hamburger-wrapper');
-  const navigation = document.querySelector('nav');
-  hamburger.addEventListener('click', () => {
-    navigation.classList.toggle('open');
+const navbase = () => {
+  const gethumburger = document.querySelector('.humburgerToggle');
+  const getclose = document.querySelector('.closeclass');
+  gethumburger.addEventListener('click', () => {
+    document.querySelector('body').classList.toggle('menu');
+  });
+  getclose.addEventListener('click', () => {
+    document.querySelector('body').classList.remove('menu');
   });
 };
 
-export default navBase;
+export default navbase;
